packages/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
dependencies = ["boto3>=1.36.16", "pydantic>=2.10.6", "fastapi>=0.115.8", "httpx>=0.28.1", "mangum>=0.19.0", "pytest-mock>=3.14.0", "requests>=2.32.3"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "NA"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
